Make graph with QQuickPaintedItem

+++ data_parser/measuredata.h
@@ -6,14 +6,14 @@
 class MeasureData
 {
 public:
-    MeasureData(const double _timeStamp, const double _value);
+    MeasureData(const long double _timeStamp, const long double _value);
 
     void printAll() const;
     QPointF getPoint() const;
 
 private:
-    double m_timeStamp;
-    double m_value;
+    long double m_timeStamp;
+    long double m_value;
 };
 
 #endif // MEASUREDATA_H
